updated map parsing, closed exit now walkable

+++ libft/ft_strlen.c
@@ -6,13 +6,13 @@
 /*   By: dominicasal <dominicasal@student.42.fr>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/16 10:41:30 by dasal             #+#    #+#             */
-/*   Updated: 2024/04/22 22:19:57 by dominicasal      ###   ########.fr       */
+/*   Updated: 2024/07/06 23:50:46 by dominicasal      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-size_t	ft_strlen(char const *str)
+int	ft_strlen(char const *str)
 {
 	size_t	i;
